+++ community/ui/header/header/header.module.scss
@@ -16,21 +16,6 @@
   }
 }
 
-.right {
-  display: flex;
-  align-items: center;
-  font-size: var(--bit-p-xs, 14px);
-  > *:not(:last-child) {
-    font-size: inherit;
-    margin-right: 8px;
-  }
-  a {
-    display: inline-flex;
-  }
-  @media screen and (max-width: $br-sm) {
-    display: none;
-  }
-}
 
 .githubLink {
   @media screen and (max-width: $br-md) {
