+++ community/ui/header/header/header.tsx
@@ -1,22 +1,42 @@
-import React, { ComponentType } from 'react';
+import React from 'react';
 import classNames from 'classnames';
-import { Header as BaseHeader } from '@teambit/design.blocks.header';
+import { useLocalStorage } from '@teambit/community.ui.hooks.use-local-storage';
+import { Icon } from '@teambit/design.elements.icon';
+import { GithubStars } from '@teambit/community.ui.github-stars';
+import { Toggle } from '@teambit/design.ui.input.toggle';
+import { ExternalLink } from '@teambit/design.ui.external-link';
+import { Header as BaseHeader, HeaderProps as BaseHeaderProps } from '@teambit/design.blocks.header';
 
 import styles from './header.module.scss';
 
-export type HeaderProps = {
-  plugins?: ComponentType[];
-} & React.HTMLAttributes<HTMLElement>;
+export type HeaderProps = {} & BaseHeaderProps;
 
 export function Header({ className, plugins, ...rest }: HeaderProps) {
+  const [highlighting, setHighlighting] = useLocalStorage('highlighting', true);
+
+  function onToggleClick(e) {
+    return setHighlighting?.(e.target.checked);
+  }
   return (
-    <BaseHeader className={classNames(className)} menuLinks={headerContent} {...rest}>
-      <div className={styles.right}>
-        {plugins?.map((Plugin, index) => (
-          <Plugin key={index} />
-        ))}
-      </div>
-    </BaseHeader>
+    <BaseHeader
+      {...rest}
+      className={classNames(className)}
+      menuLinks={headerContent}
+      plugins={[
+        () => <GithubStars className={styles.githubLink} />,
+        () => (
+          <ExternalLink href="https://join.slack.com/t/bit-dev-community/shared_invite/zt-o2tim18y-UzwOCFdTafmFKEqm2tXE4w">
+            <Icon of="slack" />
+          </ExternalLink>
+        ),
+        () => (
+          <div className={styles.inspect}>
+            <span>Inspect</span>
+            <Toggle onInputChanged={onToggleClick} checked={highlighting} />
+          </div>
+        ),
+      ]}
+    />
   );
 }
